added chain file path for failing test
